+++ src/OpenRasta.Plugins.Hydra/Configuration/HydraResourceModel.cs
@@ -1,10 +1,15 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Threading.Tasks;
+using OpenRasta.Plugins.Hydra.Internal;
 using OpenRasta.Plugins.Hydra.Schemas.Hydra;
 
 namespace OpenRasta.Plugins.Hydra.Configuration
 {
   public class HydraResourceModel
   {
+    public Func<object,SerializationOptions, Stream,Task> SerializeFunc { get; set; }
     public Vocabulary Vocabulary { get; set; }
     public List<Operation> SupportedOperations { get; set; } = new List<Operation>();
   }
