+++ src/OpenRasta.Plugins.Hydra/Internal/Serialization/SerializationOptions.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace OpenRasta.Plugins.Hydra.Internal
+{
+  public class SerializationOptions
+  {
+    public Uri BaseUri { get; set; }
+  }
+}
